Modularizing since a single cron isnt apt for handling all the coin updates with proofs.
